cmd/astrald: move mods.go to mod/
